core: Clear ambiguity between tasks and futures
A Future is an object that doesn't have a result yet and is returned
and handled by the executors used inside the core, while the core
refers to tasks as (args, kwargs) objects that bears can pass to
offload work into the core.

This commit renames some variables to reflect that difference.